pyproject.toml
[project]
name = "ai-podcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "google-api-python-client>=2.172.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.5",
    "gradio-client>=1.10.2",
    "moviepy>=2.2.1",
    "newsapi>=0.1.1",
    "newsapi-python>=0.2.7",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
]
